znalazłam literówkę i się zmotywowałam do pisania dalej

+++ src/main/java/com/kodilla/library/domain/reader/BookDto.java
@@ -9,5 +9,5 @@ public class BookDto {
     private Long id;
     private String title;
     private String author;
-    private Long publicationYear;
+    private Integer publicationYear;
 }
